COntroller for search city.
