+++ PerfectWeddingWeb/src/main/java/com/infoshareacademy/controllers/ClientController.java
@@ -7,23 +7,5 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 public class ClientController {
 
-//    @GetMapping("/")
-//    public String showCreateForm() {
-//        return "index";
-//    }
-//
-//    @GetMapping("/clients")
-//    public String clientsPage() {
-//        return "panelKlienta";
-//    }
-//
-//    @GetMapping("/providers")
-//    public String providersPage() {
-//        return "panelUslugodawcy";
-//    }
-
-
-
-
 
 }
