+++ PerfectWeddingWeb/src/main/resources/templates/ClientMenu.html
@@ -13,6 +13,8 @@
     <h2>Tu ma być wyszukiwarka</h2>
     <button onclick="window.location.href='/'">Return to Main Menu</button>
 
+
+
 </div>
 </body>
 </html>
